+++ python/tskit/tables.py
@@ -2602,3 +2602,51 @@ class TableCollection:
             self.provenances.add_row(
                 record=json.dumps(provenance.get_provenance_dict(parameters))
             )
+
+    def union(
+        self,
+        other,
+        node_mapping,
+        check_shared_equality=True,
+        add_populations=True,
+        record_provenance=True,
+    ):
+        """
+        Modifies the table collection in place by adding the non-shared
+        portions of ``other`` to itself. To perform the node-wise union,
+        the method relies on a ``node_mapping`` array, that maps nodes in
+        ``other`` to its equivalent node in ``self`` or ``tskit.NULL`` if
+        the node is exclusive to ``other``. See :meth:`TreeSequence.union` for a more
+        detailed description.
+
+        :param TableCollection other: Another table collection.
+        :param list node_mapping: An array of node IDs that relate nodes in
+            ``other`` to nodes in ``self``: the k-th element of ``node_mapping``
+            should be the index of the equivalent node in ``self``, or
+            ``tskit.NULL`` if the node is not present in ``self`` (in which case it
+            will be added to self).
+        :param bool check_shared_equality: If True, the shared portions of the
+            table collections will be checked for equality.
+        :param bool add_populations: If True, nodes new to ``self`` will be
+            assigned new population IDs.
+        :param bool record_provenance: Whether to record a provenance entry
+            in the provenance table for this operation.
+        """
+        node_mapping = util.safe_np_int_cast(node_mapping, np.int32)
+        self.ll_tables.union(
+            other.ll_tables,
+            node_mapping,
+            check_shared_equality=check_shared_equality,
+            add_populations=add_populations,
+        )
+        if record_provenance:
+            other_records = [prov.record for prov in other.provenances]
+            other_timestamps = [prov.timestamp for prov in other.provenances]
+            parameters = {
+                "command": "union",
+                "other": {"timestamp": other_timestamps, "record": other_records},
+                "node_mapping": node_mapping.tolist(),
+            }
+            self.provenances.add_row(
+                record=json.dumps(provenance.get_provenance_dict(parameters))
+            )